tasks/testsuite: move all Task-related tests over to their own file.
This is done in preparation for writing a similar bag of tests for projects.
Having them all in the same file would be a little too overwhelming.
